Add files via upload
update csv things have been moved to when you press confirm buttons. there is a new method for month combo box in scene 5 and 6, where it takes the month names and runs through a new method that is heavily commented in financeFV.java that repopulates a trends2D array (income and expense).